El tipo de dato de (x,y,z)  se cambió a float

+++ objreader/Vertex.cpp
@@ -15,7 +15,7 @@ Vertex::Vertex(){
     @param _y Coordenada en el eje y.
     @param _z Coordenada en el eje z.
 */
-Vertex::Vertex(double _x, double _y, double _z){
+Vertex::Vertex(float _x, float _y, float _z){
     x = _x;
     y = _y;
     z = _z;
@@ -39,22 +39,22 @@ void Vertex::SetId(int _id){ id = _id; }
 int Vertex::GetId(){ return id; }
 
 /*  Asigna un valor a la componente 'x' del vértice. */
-void Vertex::SetX(double _x){ x = _x; }
+void Vertex::SetX(float _x){ x = _x; }
 
 /* Devuelve el valor de la componente 'x' del vértice. */
-double Vertex::X(){ return x; }
+float Vertex::X(){ return x; }
 
 /*  Asigna un valor a la componente 'y' del vértice. */
-void Vertex::SetY(double _y){ y = _y; }
+void Vertex::SetY(float _y){ y = _y; }
 
 /* Devuelve el valor de la componente 'y' del vértice. */
-double Vertex::Y(){ return y; }
+float Vertex::Y(){ return y; }
 
 /*  Asigna un valor a la componente 'z' del vértice. */
-void Vertex::SetZ(double _z){ z = _z; }
+void Vertex::SetZ(float _z){ z = _z; }
 
 /* Devuelve el valor de la componente 'z' del vértice. */
-double Vertex::Z(){ return z; }
+float Vertex::Z(){ return z; }
 
 /*  Comparar con un vértice.
     @param _v Vértice contra el cual se compara.
